[Backend] refactor: Modify database.go structure and location
- Moved  from the  package to the global .
- Simplified the file by focusing on database initialization and connection.
- Removed specific user-related repository methods to make it a generic database utility.
